+++ src/main/kotlin/me/gurkz/superdupermod/command/SilenceMobsCommands.kt
@@ -10,58 +10,59 @@ package me.gurkz.superdupermod.command
 
 import com.mojang.brigadier.arguments.StringArgumentType
 import me.gurkz.superdupermod.config.Configs
+import me.gurkz.superdupermod.permission.KPermissions
 import net.silkmc.silk.commands.command
 import net.silkmc.silk.core.text.literalText
+import kotlin.String
 
 object SilenceMobsCommands {
-    fun registerCommands() {
-        command("superdupermod") {
-            literal("silencemobs") {
-                literal("names") {
-                    literal("list") runs {
-                        val config = Configs.superDuperConfig.silenceMobs
+    private fun config() = Configs.superDuperConfig
 
-                        val names = config.validNames.joinToString(", ")
+    fun addName(name: String): Boolean = with(config()) {
+        silenceMobs.validNames.add(name).also { if (it) save() }
+    }
 
-                        source.sendSystemMessage(literalText("current names are: $names"))
-                    }
+    fun removeName(name: String): Boolean = with(config()) {
+        silenceMobs.validNames.remove(name).also { if (it) save() }
+    }
 
-                    literal("add") {
-                        argument<String>("nameToAdd", StringArgumentType.greedyString()) { nameToAdd ->
-                            runs {
-                                val config = Configs.superDuperConfig.silenceMobs
+    fun addNameCommand() = command("superdupermod") {
+        literal("silencemobs").literal("names").literal("add").requires(KPermissions.require(
+            "superdupermod.command.silencemobs.names.add",
+            4
+        )).argument<String>("nameToAdd", StringArgumentType.greedyString()) { nameToAdd -> runs {
+            addName(nameToAdd())
 
-                                config.validNames.add(nameToAdd())
+            source.sendSystemMessage(literalText("added ${nameToAdd()}"))
+        } }
+    }
 
-                                Configs.superDuperConfig.save()
-
-                                source.sendSystemMessage(literalText("added ${nameToAdd()}"))
-
-                            }
-                        }
-                    }
-
-                    literal("remove") {
-                        argument<String>("nameToRemove", StringArgumentType.greedyString()) { nameToRemove ->
-                            suggestList { Configs.superDuperConfig.silenceMobs.validNames }
-                            runs {
-                                val config = Configs.superDuperConfig.silenceMobs
-
-                                if (!config.validNames.contains(nameToRemove())) {
-                                    source.sendFailure(literalText("that name has not been added"))
-
-                                } else {
-                                    config.validNames.remove(nameToRemove())
-
-                                    Configs.superDuperConfig.save()
-
-                                    source.sendSystemMessage(literalText("removed ${nameToRemove()}"))
-                                }
-                            }
-                        }
-                    }
+    fun removeNameCommand() = command("superdupermod") {
+        literal("silencemobs").literal("names").literal("remove").requires(KPermissions.require(
+            "superdupermod.command.silencemobs.names.remove",
+            4
+        )).argument<String>("nameToRemove", StringArgumentType.greedyString()) { nameToRemove ->
+            suggestList { config().silenceMobs.validNames }
+            runs {
+                if (removeName(nameToRemove())) {
+                    source.sendSystemMessage(literalText("removed ${nameToRemove()}"))
+                } else {
+                    source.sendFailure(literalText("that name has not been added"))
                 }
             }
         }
+    }
+
+    fun registerCommands() = command("superdupermod") {
+        literal("silencemobs").literal("names") {
+            literal("list") runs {
+                val names = config().silenceMobs.validNames.joinToString(", ")
+
+                source.sendSystemMessage(literalText("current names are: $names"))
+            }
+        }
+        addNameCommand()
+
+        removeNameCommand()
     }
 }